backend/app/services/audio_mix_service.py
import os
import uuid
import ffmpeg
from app.schemas.audio_mix import AudioMixRequest

class AudioMixService:
    """
    多轨音频合成服务
    """
    def __init__(self):
        self.output_dir = "outputs"
        os.makedirs(self.output_dir, exist_ok=True)
    
    def mix_tracks(self, request: AudioMixRequest, task_id: str):
        """
        多轨音频合成（简化版实现）
        """
        try:
            # 生成输出文件名
            output_filename = f"{task_id}.{request.output_format}"
            output_path = os.path.join(self.output_dir, output_filename)
            
            # 模拟音频处理（实际应用中这里会调用 FFmpeg）
            print(f"开始处理任务 {task_id}")
            print(f"轨道数量: {len(request.tracks)}")
            
            # 这里可以实现真正的 FFmpeg 多轨合成
            # 目前先创建一个空文件模拟完成
            with open(output_path, 'w') as f:
                f.write("Mock audio output")
            
            print(f"任务 {task_id} 处理完成，输出: {output_path}")
            return output_path
            
        except Exception as e:
            print(f"任务 {task_id} 处理失败: {str(e)}")
            raise e
